requirements.txt
PyTorch
Scipy
